monocypher: import 3.1.3
This is removing everything except the very nice compare and clear
functions, the bits of Poly1305 and Chacha we need, and their supports,
and then a couple of very minor adjustments to make it build cleanly.

I've removed it from cstyle checks because I'd rather keep it looking
like upstream as much as possible, for easier upgrading and easier
comparison.

Signed-off-by: Rob N ★ <robn@despairlabs.com>