send/recv: support for additional features in backup streams
Since we're out of DMU_BACKUP_FEATURE_* flags, we need another way to
signal required features to a reciever.

This takes the last remaining feature flag as
DMU_BACKUP_FEATURE_EXT_FEATURES, and uses it to signal that additional
features are listed in the "features" key in the nvlist in the
stream's BEGIN record.

The conception here is very simple. If the stream says it needs a
feature, then the recieving end must support that that feature in order
to recieve it. What "support" means will depend on the feature, but at
least it starts with the checks in
recv_begin_check_feature_flags_impl().