bio_map: avoid splitting ABD pages across bios
In the previous commit we prevent the kernel from inadvertently
splitting a bio in the middle of a "split page". However, there are
situations where OpenZFS itself can cause this same situation.

If we encounter a split page in the ABD, we need to add two chunks to
the bio, one for each "half" of the split. However, that requires the
bio to have two segments remaining: if there's only one, we must not add
any more to this bio; the caller (`__vdev_disk_physio()`) will then
create a new bio and we can continue to add chunks to it.

A linear ABD actually has multiple pages in a single log run. A "split"
is seen as a short "left" fragment, then a run of proper 4K pages, then
a "right" part. These are all "split", but we don't want to submit two
segments for every 4K chunk. So instead, we track when we see the "left"
part, and its size, and if we get to the end of the bio and we haven't
yet added a matching "right" part of the correct size to fix the
alignment, we compute the right amount to add to close out the bio
properly aligned.

We also make this change for scatter ABDs. Scatter chunks can be split
in the same way, though they are always added one 4K chunk at a time,
never in long runs, so the "left" and "right" parts will always be next
to one. However its theoretically possible for a gang ABD composed of a
linear and scatter ABD to finish the linear part at an "odd" segment
number, and then the scatter part will be adding pages in pairs, and end
up splitting the last page across two bios.

Signed-off-by: Rob Norris <rob.norris@klarasystems.com>
Sponsored-by: Klara, Inc.
Sponsored-by: Wasabi Technology, Inc.