zpool: add condense verb and wire up to log-spacemap flush
The idea is to have a single command that could signal to any background
cleanup task that it should do its work faster, or care less about not
getting in the way of user IO, or whatever.

This adds the the `zpool condense` command, the `ZFS_IOC_POOL_CONDENSE`
ioctl and counters so userspace can get progress. Because the target
could be anything, there's no particular unit, just a total number of
items to condense and count of how many done.

It also adds a `log-spacemap` condense target, which calls the "request
log flush" function.

Sponsored-by: Klara, Inc.
Sponsored-by: Wasabi Technology, Inc.
Signed-off-by: Rob Norris <rob.norris@klarasystems.com>