vdev_disk: make "classic" BIO submission available
The old submission code has issues, but its also been in use for a long
time. The new code should fix these things and be an overall improvement
in many areas, but does not have anywhere near the amount of testing and
real-world use to give the same confidence and understanding.

This commit adds a load-time option `zfs_vdev_disk_classic` that, when
enabled, uses the old code instead. This will allow us to backport these
changes to 2.2 and set zfs_vdev_disk_classic=1 by default, and then
recommend the option to users that need it now to fix problems or want
to try it out, while not accidentally breaking things for anyone else.

Signed-off-by: Rob Norris <rob.norris@klarasystems.com>
Sponsored-by: Klara, Inc.
Sponsored-by: Wasabi Technology, Inc.