zio: hand checksum calculation off to a taskq
Once compression and encryption are done, we don't actually need the
checksum until we actually start writing the block - DVA allocation and
other work doesn't need the checksum.

So instead of waiting for the checksum to be generated, send it out to a
background thread, and if it hasn't been generated yet by the time we're
ready to write, then we can wait for it.

TODO:
- wait at DDT_WRITE
- wait at GANG_ASSEMBLE/GANG_ISSUE
- work out how deep and wide the queue should be
- see if it actually makes writes go faster at all