kstat: replace existing kstat if name is reused
Previously, if a kstat proc name already existed, the old one would be
kept. This makes it so the old one is discarded and the new one kept.

This is not currently needed, but is aimed at future situations where
kstat setup is done during initialisation of an object, and multiple
instances of that object exist simultaneously (for example, a vdev_t
during pool import, where we both the untrusted and trusted versions
briefly exist before the untrusted one is discarded). Arguably this
shouldn't happen, but this is easier and safer than trying to rework
other codepaths to safely tear down the old kstats and bring up the new
ones.

Sponsored-by: Klara, Inc.
Sponsored-by: Syneto
Signed-off-by: Rob Norris <rob.norris@klarasystems.com>