zap_leaf: defend against chunk lists with no end marker
While it shouldn't happen, we have seen corruption where the linked list
has no end marker and the loop continues forever.

Here we add a simple counter to the loop. If we do more iterations than
max chunks in the leaf, then we've looped and we need to eject. This is
the most lightweight defense I could think of.

Sponsored-by: TrueNAS
Signed-off-by: Rob Norris <rob.norris@truenas.com>