Linux 6.15: replace has_capability() with has_capability_noaudit()
6.15 removes has_capability, as nothing uses it. Unfortunately, none of
the alternatives are easily available to us. For now, replace with
has_capability_noaudit(), which does the same thing without writing an
audit message. This is not ideal, but will do for now.

The actual use case here, not well documented, is that channel programs
capture CRED() and curproc at call time, to be rechecked when the
program runs (through the dsl_enforce_ds_ss_limits(), called from the
various async dataset ops, with their cred and proc args taken from the
asyng arg structures eg drba_cred/drba_proc).

However, the "simple" capability functions always check the credentials
of the current task, and the more complex ones that let us provide a
task or a specific credential to check are EXPORT_SYMBOL_GPL, or even
not exported at all!

Various options and moving parts for a better solution:

has_capability(proc, cap) is:

    has_ns_capability(proc, &init_user_ns, cap)

However, it's an unexported core function, so that's no good.

has_ns_capability(proc, ns, cap) is:

    rcu_read_lock();
    ret = security_capable(__task_cred(t), ns, cap, CAP_OPT_NONE);
    rcu_read_unlock();

The RCU lock is needed for __task_cred(), however we don't need it
because we already have the wanted cred. So, we could just say:

    security_capable(cr, &init_user_ns, CAP_SYS_ADMIN, CAP_OPT_NONE);

Again though, security_capable() is an unexported core function.

Incidentally, init_user_ns is EXPORT_SYMBOL_GPL, however, it's readily
accessible through init_task.init_cred.user_ns.

A cheap trick might be to use file_ns_capable(file, ns, cap) instead,
something like:

    struct file f = { .f_cred = cr };
    file_ns_capable(&f, init_task.init_cred.user_ns, CAP_SYS_ADMIN)

This is obviously abusive.

But! There is a whole thing for stashing a credential for the kernel to
use later to do some work on behalf of something else. Maybe that's
good? prepare_kernel_cred() and override_creds() and revert_creds(). I'm
not gonna try it right now, but I think we need new API to kinda
"snapshot" the current cred in zcp_eval(), and then to make it live in
the policy functions.