Linux: make migrate page/folio tests more explicit
One of the address_space_operation fields .migratepage or .migrate_folio
will always exist, depending on kernel version, however, the generic
callbacks migrate_page or migrate_folio do not exist if CONFIG_MIGRATION
is not set.

This meant that if CONFIG_MIGRATION is not set, the migrate_folio test
would fail, and we would then fall back to trying to set .migratepages,
which also didn't exist, and the build would fail.

So now we test for both possible combinations, set two separate defines,
and use them entirely based on the outcome. Since the tests include the
fallbacks, they will fail if CONFIG_MIGRATION is not set, and we won't
try to set these fields at all, which is what we want.

Sponsored-by: https://despairlabs.com/sponsor/
Signed-off-by: Rob Norris <robn@despairlabs.com>